update dev-str with user stories
Prepared 4 user stories as an example so you can all make changes and make pull request to update the others.
User stories converted to development strategies
Added `label` title to use it as a label for issues.
Split user stories in 2 categories frontend and backend.
Added "Ideas about what the page should have.." title to freely adding ideas and at the end we it can be added as an step in a user story.